Split gpu/queue.rs into focused sub-modules (#608)
Split the GPU work queue module into three focused sub-modules:
- queue/submission.rs — work item submission and batching
- queue/execution.rs — GPU execution loop and result collection
- queue/scheduling.rs — initialisation, shutdown, and Drop

Public API remains unchanged. All existing tests pass without modification.
