--- src/analysis/gpu/queue.rs
@@ -1,897 +1,0 @@
-//! GPU Work Queue module
-//!
-//! This module provides `GpuWorkQueue`, a centralised GPU thread for improved
-//! GPU utilisation when processing multiple parallel focus neurons.
-//!
-//! ## Overview (Issue #274)
-//!
-//! The `GpuWorkQueue` was introduced in v0.1.151 to solve a critical performance issue:
-//! 1. Previously each parallel focus neuron thread created its own GPU device (~100ms overhead)
-//! 2. Now a single shared queue owns the GPU thread
-//! 3. Operations are serialised through the queue, reducing device creation overhead
-//! 4. `Arc<GpuWorkQueue>` implements `GpuEvaluator` for polymorphic use
-//!
-//! ## Architecture
-//!
-//! ```text
-//! ┌─────────────────────────────────────────────────────────────┐
-//! │  CPU Threads (rayon par_iter)                               │
-//! │  ┌──────┐  ┌──────┐  ┌──────┐  ┌──────┐                    │
-//! │  │Focus1│  │Focus2│  │Focus3│  │Focus4│  ...               │
-//! │  └──┬───┘  └──┬───┘  └──┬───┘  └──┬───┘                    │
-//! │     │         │         │         │                         │
-//! │     └────┬────┴────┬────┴────┬────┘                         │
-//! │          │         │         │                              │
-//! │          ▼         ▼         ▼                              │
-//! │  ┌─────────────────────────────────────────────────┐       │
-//! │  │           GPU Work Queue (crossbeam channel)     │       │
-//! │  └──────────────────────┬──────────────────────────┘       │
-//! │                         │                                   │
-//! │                         ▼                                   │
-//! │  ┌─────────────────────────────────────────────────┐       │
-//! │  │           GPU Thread (owns GpuAnalyzer)          │       │
-//! │  │  • Batches work from multiple focus neurons      │       │
-//! │  │  • Single GPU device for all operations          │       │
-//! │  │  • Optimal GPU utilisation                       │       │
-//! │  └─────────────────────────────────────────────────┘       │
-//! └─────────────────────────────────────────────────────────────┘
-//! ```
-
-use anyhow::{Result, anyhow};
-use crossbeam_channel::{Receiver, Sender, bounded};
-use std::thread::{self, JoinHandle};
-use std::time::{Duration, Instant};
-
-use crate::analysis::gpu::analyzer::{GpuAnalyzer, GpuEvaluator};
-use crate::analysis::gpu::shaders::{GPU_INIT_TIMEOUT_SECS, GPU_SHUTDOWN_TIMEOUT_SECS};
-use crate::analysis::samples::{
-    HarmfulStats, HelpfulSample, HelpfulStats, ReluOrientation, ReluStats,
-};
-use crate::analysis::utils::{calculate_gpu_batch_timeout, get_work_queue_capacity};
-use crate::observability::{global_gpu_metrics, gpu_metrics_enabled};
-
-// =============================================================================
-// GPU Future — non-blocking handle for pending GPU results (Issue #568)
-// =============================================================================
-
-/// A handle to a pending GPU computation result.
-///
-/// Returned by `GpuWorkQueue::submit_helpful_batch` to allow the caller to
-/// perform CPU work while the GPU processes the batch, then collect the result
-/// later via `collect()`.
-pub(crate) struct GpuFuture<T> {
-    response_rx: Receiver<Result<T>>,
-    timeout: Duration,
-}
-
-impl<T> GpuFuture<T> {
-    /// Block until the GPU result is available.
-    ///
-    /// This should be called after performing any overlapping CPU work.
-    pub(crate) fn collect(self) -> Result<T> {
-        let timeout_secs = self.timeout.as_secs();
-        match self.response_rx.recv_timeout(self.timeout) {
-            Ok(result) => result,
-            Err(crossbeam_channel::RecvTimeoutError::Timeout) => Err(anyhow!(
-                "GPU batch evaluation timed out after {timeout_secs}s. \
-                 The GPU may be unresponsive. Consider reducing batch size or restarting."
-            )),
-            Err(crossbeam_channel::RecvTimeoutError::Disconnected) => {
-                Err(anyhow!("GPU response channel closed unexpectedly"))
-            }
-        }
-    }
-}
-
-// =============================================================================
-// GPU Work Queue - Centralised GPU thread for improved utilisation
-// =============================================================================
-
-/// Work request types for the GPU queue.
-/// Each variant contains the data needed for a specific GPU operation.
-pub(crate) enum GpuWorkRequest {
-    /// Batch of helpful synapse/neuron evaluations.
-    /// Each item is a slice of samples to evaluate.
-    HelpfulBatch {
-        /// Samples for each evaluation, indexed by request ID.
-        samples: Vec<Vec<HelpfulSample>>,
-        /// Channel to send results back.
-        response_tx: Sender<Result<Vec<HelpfulStats>>>,
-    },
-    /// Batch of harmful synapse evaluations.
-    /// Each item is (samples, weight) pair.
-    HarmfulBatch {
-        /// (samples, weight) pairs for each evaluation.
-        samples_with_weights: Vec<(Vec<HelpfulSample>, f32)>,
-        /// Channel to send results back.
-        response_tx: Sender<Result<Vec<HarmfulStats>>>,
-    },
-    /// ReLU activation evaluation for neuron candidates.
-    ReluEval {
-        samples: Vec<HelpfulSample>,
-        threshold: f32,
-        response_tx: Sender<Result<(ReluStats, ReluStats, f32)>>,
-    },
-    /// General activation function evaluation for neuron candidates.
-    ActivationEval {
-        samples: Vec<HelpfulSample>,
-        activation_type: u32,
-        orientation: f32,
-        scale: f32,
-        response_tx: Sender<Result<(f32, f32, f32, u32)>>,
-    },
-    /// Batched activation function evaluation for multiple configs.
-    /// Issue #201: Reduces GPU round-trips by evaluating multiple activation
-    /// configurations in a single command buffer submission.
-    #[allow(clippy::type_complexity)]
-    ActivationBatchEval {
-        samples: Vec<HelpfulSample>,
-        activation_configs: Vec<(u32, f32, f32)>, // (activation_type, orientation, scale)
-        response_tx: Sender<Result<Vec<(f32, f32, f32, u32)>>>,
-    },
-    /// Request to shut down the GPU thread.
-    Shutdown,
-}
-
-/// Centralised GPU work queue that processes all GPU operations on a single thread.
-///
-/// This eliminates the overhead of creating multiple GPU devices (one per parallel
-/// focus neuron) and improves GPU utilisation by batching work from multiple sources.
-pub struct GpuWorkQueue {
-    /// Channel to send work to the GPU thread.
-    work_tx: Sender<GpuWorkRequest>,
-    /// Handle to the GPU thread (for clean shutdown).
-    thread_handle: Option<JoinHandle<()>>,
-    /// Channel to receive notification when GPU thread exits.
-    /// This allows Drop to use a timeout instead of blocking forever.
-    exit_rx: Receiver<()>,
-}
-
-impl GpuWorkQueue {
-    /// Create a new GPU work queue with a dedicated GPU thread.
-    ///
-    /// The GPU thread is spawned immediately and owns the GpuAnalyzer.
-    /// All GPU operations are processed sequentially on this thread,
-    /// eliminating device creation overhead and improving utilisation.
-    ///
-    /// CRITICAL: The GpuAnalyzer is created INSIDE the GPU thread, not before.
-    /// wgpu devices have thread-local state that doesn't transfer properly when
-    /// moved across threads, causing deadlocks in device.poll().
-    pub fn new() -> Result<Self> {
-        // Create the channel for sending work to the GPU thread.
-        // Capacity is dynamically sized based on available system memory.
-        // Lower capacity = more backpressure = less memory usage.
-        // This prevents Metal command buffer exhaustion on memory-constrained systems.
-        let queue_capacity = get_work_queue_capacity();
-        let (work_tx, work_rx): (Sender<GpuWorkRequest>, Receiver<GpuWorkRequest>) =
-            bounded(queue_capacity);
-
-        // Channel to receive initialisation result from the GPU thread.
-        // This ensures the GpuAnalyzer is created ON the GPU thread, not moved to it.
-        let (init_tx, init_rx): (Sender<Result<()>>, Receiver<Result<()>>) = bounded(1);
-
-        // Channel to receive notification when GPU thread exits.
-        // This allows Drop to use a timeout instead of blocking forever if the GPU hangs.
-        let (exit_tx, exit_rx): (Sender<()>, Receiver<()>) = bounded(1);
-
-        // Spawn dedicated GPU thread - analyzer is created INSIDE this thread
-        let thread_handle = thread::spawn(move || {
-            // Create analyzer on THIS thread to avoid wgpu thread-local state issues
-            match GpuAnalyzer::new() {
-                Ok(analyzer) => {
-                    // Signal successful initialisation
-                    let _ = init_tx.send(Ok(()));
-                    // Run the main loop
-                    Self::gpu_thread_loop(analyzer, work_rx);
-                }
-                Err(e) => {
-                    // Signal initialisation failure
-                    let _ = init_tx.send(Err(e));
-                }
-            }
-            // Always signal exit, even if initialisation failed or loop panicked
-            let _ = exit_tx.send(());
-        });
-
-        // Wait for initialisation to complete with timeout
-        let init_timeout = Duration::from_secs(GPU_INIT_TIMEOUT_SECS);
-        match init_rx.recv_timeout(init_timeout) {
-            Ok(Ok(())) => {} // Success
-            Ok(Err(e)) => return Err(e),
-            Err(crossbeam_channel::RecvTimeoutError::Timeout) => {
-                return Err(anyhow!(
-                    "GPU initialisation timed out after {GPU_INIT_TIMEOUT_SECS}s. \
-                     The GPU may be unresponsive or overwhelmed. \
-                     Try restarting the process or reducing workload."
-                ));
-            }
-            Err(crossbeam_channel::RecvTimeoutError::Disconnected) => {
-                return Err(anyhow!("GPU thread failed to start (channel disconnected)"));
-            }
-        }
-
-        Ok(Self {
-            work_tx,
-            thread_handle: Some(thread_handle),
-            exit_rx,
-        })
-    }
-
-    /// The main loop for the GPU thread.
-    /// Processes work requests until shutdown is requested.
-    ///
-    /// When `NEAT_AI_DISCOVERY_GPU_METRICS=1` is set, this loop tracks:
-    /// - Batch count and samples processed
-    /// - GPU busy time (time spent executing GPU operations)
-    fn gpu_thread_loop(analyzer: GpuAnalyzer, work_rx: Receiver<GpuWorkRequest>) {
-        let track_metrics = gpu_metrics_enabled();
-
-        while let Ok(request) = work_rx.recv() {
-            match request {
-                GpuWorkRequest::HelpfulBatch {
-                    samples,
-                    response_tx,
-                } => {
-                    let sample_count: usize = samples.iter().map(|v| v.len()).sum();
-
-                    let start = if track_metrics {
-                        Some(Instant::now())
-                    } else {
-                        None
-                    };
-
-                    // Convert Vec<Vec<HelpfulSample>> to &[&[HelpfulSample]] for the API
-                    let samples_refs: Vec<&[HelpfulSample]> =
-                        samples.iter().map(|v| v.as_slice()).collect();
-                    let result = analyzer.evaluate_helpful_batch(&samples_refs);
-
-                    if let Some(start) = start {
-                        let metrics = global_gpu_metrics();
-                        metrics.record_batch(sample_count);
-                        metrics.record_gpu_busy_us(start.elapsed().as_micros() as u64);
-                    }
-
-                    // Send result back (ignore send errors - receiver may have dropped)
-                    let _ = response_tx.send(result);
-                }
-                GpuWorkRequest::HarmfulBatch {
-                    samples_with_weights,
-                    response_tx,
-                } => {
-                    let sample_count: usize =
-                        samples_with_weights.iter().map(|(v, _)| v.len()).sum();
-
-                    let start = if track_metrics {
-                        Some(Instant::now())
-                    } else {
-                        None
-                    };
-
-                    // Convert to the format expected by evaluate_harmful_batch
-                    let batch_refs: Vec<(&[HelpfulSample], f32)> = samples_with_weights
-                        .iter()
-                        .map(|(samples, weight)| (samples.as_slice(), *weight))
-                        .collect();
-                    let result = analyzer.evaluate_harmful_batch(&batch_refs);
-
-                    if let Some(start) = start {
-                        let metrics = global_gpu_metrics();
-                        metrics.record_batch(sample_count);
-                        metrics.record_gpu_busy_us(start.elapsed().as_micros() as u64);
-                    }
-
-                    let _ = response_tx.send(result);
-                }
-                GpuWorkRequest::ReluEval {
-                    samples,
-                    threshold,
-                    response_tx,
-                } => {
-                    let sample_count = samples.len();
-
-                    let start = if track_metrics {
-                        Some(Instant::now())
-                    } else {
-                        None
-                    };
-
-                    let result = analyzer.evaluate_relu_gpu(&samples, threshold);
-
-                    if let Some(start) = start {
-                        let metrics = global_gpu_metrics();
-                        metrics.record_batch(sample_count);
-                        metrics.record_gpu_busy_us(start.elapsed().as_micros() as u64);
-                    }
-
-                    let _ = response_tx.send(result);
-                }
-                GpuWorkRequest::ActivationEval {
-                    samples,
-                    activation_type,
-                    orientation,
-                    scale,
-                    response_tx,
-                } => {
-                    let sample_count = samples.len();
-
-                    let start = if track_metrics {
-                        Some(Instant::now())
-                    } else {
-                        None
-                    };
-
-                    let result = analyzer.evaluate_activation_gpu(
-                        &samples,
-                        activation_type,
-                        orientation,
-                        scale,
-                    );
-
-                    if let Some(start) = start {
-                        let metrics = global_gpu_metrics();
-                        metrics.record_batch(sample_count);
-                        metrics.record_gpu_busy_us(start.elapsed().as_micros() as u64);
-                    }
-
-                    let _ = response_tx.send(result);
-                }
-                GpuWorkRequest::ActivationBatchEval {
-                    samples,
-                    activation_configs,
-                    response_tx,
-                } => {
-                    // Issue #201: Batched activation evaluation
-                    let sample_count = samples.len();
-                    let config_count = activation_configs.len();
-
-                    let start = if track_metrics {
-                        Some(Instant::now())
-                    } else {
-                        None
-                    };
-
-                    let result =
-                        analyzer.evaluate_activations_batched_gpu(&samples, &activation_configs);
-
-                    if let Some(start) = start {
-                        let metrics = global_gpu_metrics();
-                        // Record total samples processed (samples * configs)
-                        metrics.record_batch(sample_count * config_count);
-                        metrics.record_gpu_busy_us(start.elapsed().as_micros() as u64);
-                    }
-
-                    let _ = response_tx.send(result);
-                }
-                GpuWorkRequest::Shutdown => {
-                    // Clean shutdown requested
-                    break;
-                }
-            }
-        }
-    }
-
-    /// Submit a helpful batch to the GPU thread without blocking (Issue #568).
-    ///
-    /// Returns a `GpuFuture` that can be collected later, allowing the caller
-    /// to perform CPU work (e.g., preparing harmful samples) while the GPU
-    /// processes the helpful batch.
-    pub(crate) fn submit_helpful_batch(
-        &self,
-        samples: Vec<Vec<HelpfulSample>>,
-        deadline: &Option<std::time::SystemTime>,
-    ) -> Result<GpuFuture<Vec<HelpfulStats>>> {
-        if samples.is_empty() {
-            // Return a pre-resolved future with empty results
-            let (tx, rx) = bounded(1);
-            let _ = tx.send(Ok(Vec::new()));
-            return Ok(GpuFuture {
-                response_rx: rx,
-                timeout: Duration::from_secs(1),
-            });
-        }
-
-        let (response_tx, response_rx) = bounded(1);
-        let timeout = calculate_gpu_batch_timeout(deadline);
-        let timeout_secs = timeout.as_secs();
-
-        match self.work_tx.send_timeout(
-            GpuWorkRequest::HelpfulBatch {
-                samples,
-                response_tx,
-            },
-            timeout,
-        ) {
-            Ok(()) => {}
-            Err(crossbeam_channel::SendTimeoutError::Timeout(_)) => {
-                return Err(anyhow!(
-                    "GPU work queue full - send timed out after {timeout_secs}s. \
-                     The GPU thread may be hung. Consider restarting the process."
-                ));
-            }
-            Err(crossbeam_channel::SendTimeoutError::Disconnected(_)) => {
-                return Err(anyhow!("GPU work queue channel closed"));
-            }
-        }
-
-        Ok(GpuFuture {
-            response_rx,
-            timeout,
-        })
-    }
-
-    /// Submit a batch of helpful evaluations and wait for results.
-    ///
-    /// This is a synchronous call that blocks until the GPU thread processes
-    /// the batch and returns results.
-    ///
-    /// The `deadline` parameter is used to calculate an adaptive timeout:
-    /// - With deadline: uses up to half remaining time (60s-5min)
-    /// - Without deadline: uses maximum timeout (5 minutes)
-    pub fn evaluate_helpful_batch(
-        &self,
-        samples: Vec<Vec<HelpfulSample>>,
-        deadline: &Option<std::time::SystemTime>,
-    ) -> Result<Vec<HelpfulStats>> {
-        if samples.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        // Create a one-shot channel for the response
-        let (response_tx, response_rx) = bounded(1);
-
-        // Calculate timeout based on remaining deadline
-        let timeout = calculate_gpu_batch_timeout(deadline);
-        let timeout_secs = timeout.as_secs();
-
-        // Send the work request with timeout to prevent deadlock if GPU thread is hung
-        // If the channel is full (GPU not processing), this will timeout instead of blocking forever
-        match self.work_tx.send_timeout(
-            GpuWorkRequest::HelpfulBatch {
-                samples,
-                response_tx,
-            },
-            timeout,
-        ) {
-            Ok(()) => {}
-            Err(crossbeam_channel::SendTimeoutError::Timeout(_)) => {
-                return Err(anyhow!(
-                    "GPU work queue full - send timed out after {timeout_secs}s. \
-                     The GPU thread may be hung. Consider restarting the process."
-                ));
-            }
-            Err(crossbeam_channel::SendTimeoutError::Disconnected(_)) => {
-                return Err(anyhow!("GPU work queue channel closed"));
-            }
-        }
-
-        // Wait for the response with timeout
-        match response_rx.recv_timeout(timeout) {
-            Ok(result) => result,
-            Err(crossbeam_channel::RecvTimeoutError::Timeout) => Err(anyhow!(
-                "GPU helpful batch evaluation timed out after {timeout_secs}s. \
-                     The GPU may be unresponsive. Consider reducing batch size or restarting."
-            )),
-            Err(crossbeam_channel::RecvTimeoutError::Disconnected) => {
-                Err(anyhow!("GPU response channel closed unexpectedly"))
-            }
-        }
-    }
-
-    /// Submit a batch of harmful evaluations and wait for results.
-    ///
-    /// The `deadline` parameter is used to calculate an adaptive timeout (60s-5min).
-    pub fn evaluate_harmful_batch(
-        &self,
-        samples_with_weights: Vec<(Vec<HelpfulSample>, f32)>,
-        deadline: &Option<std::time::SystemTime>,
-    ) -> Result<Vec<HarmfulStats>> {
-        if samples_with_weights.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        let (response_tx, response_rx) = bounded(1);
-        let timeout = calculate_gpu_batch_timeout(deadline);
-        let timeout_secs = timeout.as_secs();
-
-        // Send with timeout to prevent deadlock if GPU thread is hung
-        match self.work_tx.send_timeout(
-            GpuWorkRequest::HarmfulBatch {
-                samples_with_weights,
-                response_tx,
-            },
-            timeout,
-        ) {
-            Ok(()) => {}
-            Err(crossbeam_channel::SendTimeoutError::Timeout(_)) => {
-                return Err(anyhow!(
-                    "GPU work queue full - send timed out after {timeout_secs}s. \
-                     The GPU thread may be hung. Consider restarting the process."
-                ));
-            }
-            Err(crossbeam_channel::SendTimeoutError::Disconnected(_)) => {
-                return Err(anyhow!("GPU work queue channel closed"));
-            }
-        }
-
-        match response_rx.recv_timeout(timeout) {
-            Ok(result) => result,
-            Err(crossbeam_channel::RecvTimeoutError::Timeout) => Err(anyhow!(
-                "GPU harmful batch evaluation timed out after {timeout_secs}s. \
-                     The GPU may be unresponsive. Consider reducing batch size or restarting."
-            )),
-            Err(crossbeam_channel::RecvTimeoutError::Disconnected) => {
-                Err(anyhow!("GPU response channel closed unexpectedly"))
-            }
-        }
-    }
-
-    /// Submit a ReLU evaluation and wait for results.
-    /// Returns (positive_stats, negative_stats, baseline_error_sq).
-    ///
-    /// The `deadline` parameter is used to calculate an adaptive timeout (60s-5min).
-    pub(crate) fn evaluate_relu_gpu(
-        &self,
-        samples: Vec<HelpfulSample>,
-        threshold: f32,
-        deadline: &Option<std::time::SystemTime>,
-    ) -> Result<(ReluStats, ReluStats, f32)> {
-        if samples.is_empty() {
-            return Ok((
-                ReluStats::new(ReluOrientation::Positive),
-                ReluStats::new(ReluOrientation::Negative),
-                0.0,
-            ));
-        }
-
-        let (response_tx, response_rx) = bounded(1);
-        let timeout = calculate_gpu_batch_timeout(deadline);
-        let timeout_secs = timeout.as_secs();
-
-        // Send with timeout to prevent deadlock if GPU thread is hung
-        match self.work_tx.send_timeout(
-            GpuWorkRequest::ReluEval {
-                samples,
-                threshold,
-                response_tx,
-            },
-            timeout,
-        ) {
-            Ok(()) => {}
-            Err(crossbeam_channel::SendTimeoutError::Timeout(_)) => {
-                return Err(anyhow!(
-                    "GPU work queue full - send timed out after {timeout_secs}s. \
-                     The GPU thread may be hung. Consider restarting the process."
-                ));
-            }
-            Err(crossbeam_channel::SendTimeoutError::Disconnected(_)) => {
-                return Err(anyhow!("GPU work queue channel closed"));
-            }
-        }
-
-        match response_rx.recv_timeout(timeout) {
-            Ok(result) => result,
-            Err(crossbeam_channel::RecvTimeoutError::Timeout) => Err(anyhow!(
-                "GPU ReLU evaluation timed out after {timeout_secs}s. \
-                     The GPU may be unresponsive. Consider reducing batch size or restarting."
-            )),
-            Err(crossbeam_channel::RecvTimeoutError::Disconnected) => {
-                Err(anyhow!("GPU response channel closed unexpectedly"))
-            }
-        }
-    }
-
-    /// Submit an activation evaluation and wait for results.
-    /// Returns (sum_activation_sq, sum_error_activation, total_baseline_error_sq, improved_count).
-    ///
-    /// The `deadline` parameter is used to calculate an adaptive timeout (60s-5min).
-    pub(crate) fn evaluate_activation_gpu(
-        &self,
-        samples: Vec<HelpfulSample>,
-        activation_type: u32,
-        orientation: f32,
-        scale: f32,
-        deadline: &Option<std::time::SystemTime>,
-    ) -> Result<(f32, f32, f32, u32)> {
-        if samples.is_empty() {
-            return Ok((0.0, 0.0, 0.0, 0));
-        }
-
-        let (response_tx, response_rx) = bounded(1);
-        let timeout = calculate_gpu_batch_timeout(deadline);
-        let timeout_secs = timeout.as_secs();
-
-        // Send with timeout to prevent deadlock if GPU thread is hung
-        match self.work_tx.send_timeout(
-            GpuWorkRequest::ActivationEval {
-                samples,
-                activation_type,
-                orientation,
-                scale,
-                response_tx,
-            },
-            timeout,
-        ) {
-            Ok(()) => {}
-            Err(crossbeam_channel::SendTimeoutError::Timeout(_)) => {
-                return Err(anyhow!(
-                    "GPU work queue full - send timed out after {timeout_secs}s. \
-                     The GPU thread may be hung. Consider restarting the process."
-                ));
-            }
-            Err(crossbeam_channel::SendTimeoutError::Disconnected(_)) => {
-                return Err(anyhow!("GPU work queue channel closed"));
-            }
-        }
-
-        match response_rx.recv_timeout(timeout) {
-            Ok(result) => result,
-            Err(crossbeam_channel::RecvTimeoutError::Timeout) => Err(anyhow!(
-                "GPU activation evaluation timed out after {timeout_secs}s. \
-                     The GPU may be unresponsive. Consider reducing batch size or restarting."
-            )),
-            Err(crossbeam_channel::RecvTimeoutError::Disconnected) => {
-                Err(anyhow!("GPU response channel closed unexpectedly"))
-            }
-        }
-    }
-
-    /// Submit a batched activation evaluation and wait for results.
-    ///
-    /// Issue #201: Evaluates multiple activation function configurations in a single
-    /// GPU command buffer submission, reducing CPU-GPU round-trips by 10-20%.
-    ///
-    /// # Arguments
-    /// * `samples` - The sample data to evaluate
-    /// * `activation_configs` - List of (activation_type, orientation, scale) tuples
-    /// * `deadline` - Optional deadline for adaptive timeout calculation
-    ///
-    /// # Returns
-    /// Vector of (sum_activation_sq, sum_error_activation, total_baseline_error_sq, improved_count)
-    /// in the same order as the input configs.
-    pub(crate) fn evaluate_activations_batched_gpu(
-        &self,
-        samples: Vec<HelpfulSample>,
-        activation_configs: Vec<(u32, f32, f32)>,
-        deadline: &Option<std::time::SystemTime>,
-    ) -> Result<Vec<(f32, f32, f32, u32)>> {
-        // Handle edge cases
-        if activation_configs.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        if samples.is_empty() {
-            // Return zero results for each config
-            return Ok(vec![(0.0, 0.0, 0.0, 0); activation_configs.len()]);
-        }
-
-        let (response_tx, response_rx) = bounded(1);
-        let timeout = calculate_gpu_batch_timeout(deadline);
-        let timeout_secs = timeout.as_secs();
-
-        // Send with timeout to prevent deadlock if GPU thread is hung
-        match self.work_tx.send_timeout(
-            GpuWorkRequest::ActivationBatchEval {
-                samples,
-                activation_configs,
-                response_tx,
-            },
-            timeout,
-        ) {
-            Ok(()) => {}
-            Err(crossbeam_channel::SendTimeoutError::Timeout(_)) => {
-                return Err(anyhow!(
-                    "GPU work queue full - send timed out after {timeout_secs}s. \
-                     The GPU thread may be hung. Consider restarting the process."
-                ));
-            }
-            Err(crossbeam_channel::SendTimeoutError::Disconnected(_)) => {
-                return Err(anyhow!("GPU work queue channel closed"));
-            }
-        }
-
-        match response_rx.recv_timeout(timeout) {
-            Ok(result) => result,
-            Err(crossbeam_channel::RecvTimeoutError::Timeout) => Err(anyhow!(
-                "GPU batched activation evaluation timed out after {timeout_secs}s. \
-                     The GPU may be unresponsive. Consider reducing batch size or restarting."
-            )),
-            Err(crossbeam_channel::RecvTimeoutError::Disconnected) => {
-                Err(anyhow!("GPU response channel closed unexpectedly"))
-            }
-        }
-    }
-
-    /// Request the GPU thread to shut down.
-    /// This should be called before dropping the queue to ensure clean shutdown.
-    ///
-    /// Uses a timeout to avoid blocking forever if the queue is full and the GPU
-    /// thread is hung. If the send times out, the GPU thread is likely unresponsive
-    /// and the Drop implementation will handle cleanup via the exit_rx timeout.
-    pub fn shutdown(&self) {
-        // Use timeout to avoid blocking forever if queue is full and GPU thread is hung.
-        // 2 seconds is generous - if the GPU thread is responsive, it should drain
-        // items much faster. If this times out, proceed to exit_rx timeout in Drop.
-        let shutdown_send_timeout = Duration::from_secs(2);
-        let _ = self
-            .work_tx
-            .send_timeout(GpuWorkRequest::Shutdown, shutdown_send_timeout);
-    }
-}
-
-// Note: GPU_SHUTDOWN_TIMEOUT_SECS is imported from gpu/shaders.rs (Issue #277)
-
-impl Drop for GpuWorkQueue {
-    fn drop(&mut self) {
-        // Request shutdown
-        self.shutdown();
-
-        // Wait for GPU thread to exit with timeout
-        // This prevents hanging forever if the GPU driver is stuck (e.g., Metal semaphore wait)
-        let timeout = Duration::from_secs(GPU_SHUTDOWN_TIMEOUT_SECS);
-        match self.exit_rx.recv_timeout(timeout) {
-            Ok(()) => {
-                // Thread exited cleanly, now safe to join
-                if let Some(handle) = self.thread_handle.take() {
-                    let _ = handle.join();
-                }
-            }
-            Err(crossbeam_channel::RecvTimeoutError::Timeout) => {
-                // GPU thread is stuck (likely in Metal driver)
-                // Log warning and abandon the thread - it will be cleaned up on process exit
-                tracing::warn!(
-                    timeout_secs = GPU_SHUTDOWN_TIMEOUT_SECS,
-                    "GPU thread did not exit within timeout — the GPU driver may be hung, \
-                     abandoning thread to prevent deadlock. Consider restarting the process."
-                );
-                // Don't join - the thread is stuck and joining would block forever
-                let _ = self.thread_handle.take();
-            }
-            Err(crossbeam_channel::RecvTimeoutError::Disconnected) => {
-                // Channel disconnected - thread already exited (possibly via panic)
-                if let Some(handle) = self.thread_handle.take() {
-                    let _ = handle.join();
-                }
-            }
-        }
-    }
-}
-
-/// Implementation for shared GpuWorkQueue.
-/// NOTE: These trait methods use `None` deadline, which gives maximum timeout (5 minutes).
-/// For deadline-aware evaluation, use the batch methods directly with an explicit deadline.
-impl GpuEvaluator for GpuWorkQueue {
-    fn evaluate_relu(
-        &self,
-        samples: &[HelpfulSample],
-        threshold: f32,
-    ) -> Result<(ReluStats, ReluStats, f32)> {
-        // Clone samples to send to the GPU thread
-        // Uses None deadline = maximum timeout (5 minutes)
-        self.evaluate_relu_gpu(samples.to_vec(), threshold, &None)
-    }
-
-    fn evaluate_activation(
-        &self,
-        samples: &[HelpfulSample],
-        activation_type: u32,
-        orientation: f32,
-        scale: f32,
-    ) -> Result<(f32, f32, f32, u32)> {
-        // Uses None deadline = maximum timeout (5 minutes)
-        self.evaluate_activation_gpu(samples.to_vec(), activation_type, orientation, scale, &None)
-    }
-
-    fn evaluate_activations_batched(
-        &self,
-        samples: &[HelpfulSample],
-        activation_configs: &[(u32, f32, f32)],
-    ) -> Result<Vec<(f32, f32, f32, u32)>> {
-        // Uses None deadline = maximum timeout (5 minutes)
-        self.evaluate_activations_batched_gpu(samples.to_vec(), activation_configs.to_vec(), &None)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Test that the GpuWorkQueue module exports all expected types.
-    #[test]
-    fn test_queue_module_exports_are_accessible() {
-        // Verify that GpuWorkQueue struct is accessible
-        fn _takes_queue(_: &GpuWorkQueue) {}
-
-        // Verify GpuWorkRequest enum is accessible (pub(crate))
-        // This is tested implicitly through the GpuWorkQueue methods
-    }
-
-    /// Test that GpuWorkQueue implements GpuEvaluator trait.
-    #[test]
-    fn test_gpu_work_queue_implements_evaluator() {
-        // Verify GpuEvaluator trait is implemented
-        fn _takes_evaluator<T: GpuEvaluator>(_: &T) {}
-
-        // We can't create an instance without GPU, but we can verify the trait bound
-        fn _verify_impl() {
-            fn _dummy(_queue: &GpuWorkQueue) {
-                _takes_evaluator(_queue);
-            }
-        }
-    }
-
-    /// Test that empty batch returns empty results without GPU.
-    #[test]
-    fn test_empty_helpful_batch_returns_empty_vec() {
-        // We can't test this without creating a GpuWorkQueue (requires GPU),
-        // but we can verify the early return logic exists by examining the code.
-        // The actual test is that the code compiles with the correct return type.
-        let empty: Vec<Vec<HelpfulSample>> = Vec::new();
-        assert!(empty.is_empty());
-    }
-
-    /// Test that empty harmful batch returns empty results without GPU.
-    #[test]
-    fn test_empty_harmful_batch_returns_empty_vec() {
-        let empty: Vec<(Vec<HelpfulSample>, f32)> = Vec::new();
-        assert!(empty.is_empty());
-    }
-
-    /// Test that empty ReLU samples returns default stats without GPU.
-    #[test]
-    fn test_empty_relu_samples_returns_defaults() {
-        let empty: Vec<HelpfulSample> = Vec::new();
-        assert!(empty.is_empty());
-    }
-
-    /// Test that empty activation samples returns default stats without GPU.
-    #[test]
-    fn test_empty_activation_samples_returns_defaults() {
-        let empty: Vec<HelpfulSample> = Vec::new();
-        assert!(empty.is_empty());
-    }
-
-    /// Test that the shutdown timeout constant is reasonable.
-    #[test]
-    fn test_shutdown_timeout_is_reasonable() {
-        // Shutdown timeout should be long enough to allow graceful shutdown
-        // but not so long that it hangs the process indefinitely
-        // Using const assertion to verify at compile time
-        const _: () = assert!(GPU_SHUTDOWN_TIMEOUT_SECS >= 5, "Shutdown timeout too short");
-        const _: () = assert!(GPU_SHUTDOWN_TIMEOUT_SECS <= 30, "Shutdown timeout too long");
-    }
-
-    /// Test that GpuWorkRequest variants can be constructed (compile-time check).
-    #[test]
-    fn test_gpu_work_request_variants_constructible() {
-        // Verify all variants can be constructed
-        let (tx, _rx) = bounded::<Result<Vec<HelpfulStats>>>(1);
-        let _helpful = GpuWorkRequest::HelpfulBatch {
-            samples: vec![],
-            response_tx: tx,
-        };
-
-        let (tx, _rx) = bounded::<Result<Vec<HarmfulStats>>>(1);
-        let _harmful = GpuWorkRequest::HarmfulBatch {
-            samples_with_weights: vec![],
-            response_tx: tx,
-        };
-
-        let (tx, _rx) = bounded::<Result<(ReluStats, ReluStats, f32)>>(1);
-        let _relu = GpuWorkRequest::ReluEval {
-            samples: vec![],
-            threshold: 0.0,
-            response_tx: tx,
-        };
-
-        let (tx, _rx) = bounded::<Result<(f32, f32, f32, u32)>>(1);
-        let _activation = GpuWorkRequest::ActivationEval {
-            samples: vec![],
-            activation_type: 0,
-            orientation: 1.0,
-            scale: 1.0,
-            response_tx: tx,
-        };
-
-        let _shutdown = GpuWorkRequest::Shutdown;
-    }
-}
